requirements.txt
# Core dependencies
faster-whisper>=1.0.0
transformers>=4.36.0
torch>=2.1.0
torchaudio>=2.1.0
sentencepiece>=0.1.99

# Audio processing
sounddevice>=0.4.6
numpy>=1.24.0
scipy>=1.11.0

# Voice activity detection
silero-vad>=5.0

# Text-to-speech
piper-tts>=1.2.0

# Web server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
websockets>=12.0

# Utilities
pyyaml>=6.0.1
python-multipart>=0.0.6
zeroconf
